refactor(query-service): fold test fixtures into the provider tests and trim comments
