+++ hotshot-query-service/src/fetching/provider.rs
@@ -43,8 +43,6 @@ use super::Request;
 
 mod any;
 mod query_service;
-#[cfg(test)]
-pub(crate) mod test_fixtures;
 mod testing;
 
 pub use any::AnyProvider;
